+++ module/item/item.js
@@ -1,5 +1,8 @@
 import { CY } from "../config.js";
 import { drawDocument, dupeData } from "../packutils.js";
+import { soundEffects } from "../settings.js";
+import { uiEject, uiError, uiSlot } from "../sound.js";
+import { byName, rollTotal } from "../utils.js";
 
 /**
  * @extends {Item}
@@ -59,11 +62,62 @@ import { drawDocument, dupeData } from "../packutils.js";
   /** @override */
   prepareDerivedData() {
     super.prepareDerivedData();
-    if (this.system.nanoId) {
+    if (this.type === CY.itemTypes.app && this.parent) {
+    } else if (this.type === CY.itemTypes.cyberdeck && this.parent) {
+      this.system.slottedAppsData = this.slottedApps();
+      const rollData = this.parent.getRollData();
+      if (this.system.slotFormula) {
+        this.system.slots = rollTotal(this.system.slotFormula, rollData);
+      } else {
+        this.system.slots = 1;
+      }
+      this.system.slotsUsed = this.slottedApps().length;
+    } else if (this.type === CY.itemTypes.infestation && this.system.nanoId) {
       this.system.nanoName = this.linkedNano()?.name;
-    }
-    if (this.system.infestationId) {
+    } else if (this.type === CY.itemTypes.nanoPower && this.system.infestationId) {
       this.system.infestationName = this.linkedInfestation()?.name;
     }
+  }
+
+  // TODO:L figure out if we need this, or if using this.actor in prepareDerivedData() suffices
+  // prepareActorDerivedData(actor) {
+  // }
+
+  slottedApps() {
+    return this.parent?.items
+      .filter(item => item.type === CY.itemTypes.app)
+      .filter(item => item.system.cyberdeckId === this._id)
+      .sort(byName);
+  }  
+
+  async slotApp(app) {
+    if (app.system.cyberdeckId === this._id) {
+      // already slotted in this deck
+      return;
+    }
+    if (this.slottedApps().length < this.system.slots) {
+      // slots available, so slot it
+      if (soundEffects()) {
+        uiSlot();
+        setTimeout(() => app.update({ "data.cyberdeckId": this._id }), 1000);  
+      } else {
+        await app.update({ "data.cyberdeckId": this._id });
+      }
+    } else {
+      // no empty slots
+      ui.notifications.error(game.i18n.localize('CY.NoEmptySlots'));
+      uiError();
+    }
+  }
+
+  async eject() {
+    if (this.system.cyberdeckId) {
+      if (soundEffects()) {
+        uiEject();
+        setTimeout(() => this.update({ "data.cyberdeckId": null }), 1000);
+      } else {
+        await this.update({ "data.cyberdeckId": null });
+      }
+    }        
   }
 }